AmPac-App-main/apps/brain/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
httpx==0.26.0
python-multipart==0.0.9
msgraph-sdk==1.0.0
azure-identity==1.15.0
pydantic-settings==2.1.0
stripe==8.0.0
PyJWT==2.8.0
groq==0.4.1
pdfplumber==0.10.3
python-docx==1.1.0
sentry-sdk[fastapi]==1.39.1
